--- homeworks/week21/hw1/.eslintrc.json
@@ -1,3 +1,0 @@
-{
-  "extends": ["react-app"]
-}
