--- homeworks/week21/hw1/src/styles.css
@@ -1,12 +1,0 @@
-html,
-body {
-  margin: 0;
-  padding: 0;
-  height: 100%;
-  width: 100%;
-}
-
-#root {
-  height: 100%;
-  width: 100%;
-}
